Split away Profile types, Summary and Edit components. Initial POST Profile implementation
